Preface militia power with acknowledgement of danger of standing army
Col: Mason, being sensible that an absolute prohibition of standing armies in time of peace might be unsafe, and wishing at the same time to insert something pointing out and guarding against the danger of them, moved to preface the clause (Art I sect. 8) "To provide for organizing, arming and disciplining the Militia &c" with the words "And that the liberties of the people may be better secured against the danger of standing armies in time of peace" Mr. Randolph 2ded. the motion Mr Madison was in favor of it. It did not restrain Congress from establishing a military force in time of peace if found necessary; and as armies in time of peace are allowed on all hands to be an evil, it is well to discountenance them by the Constitution, as far as will consist with the essential power of the Govt. on that head.