Add "or other direct tax" to avoid loopholes
Mr Read moved to insert after "capitation" the words. "or other direct tax" He was afraid that some liberty might otherwise be taken to saddle the States with a readjustment by this rule, of past Requisitions of Congs - and that his amendment by giving another cast to the meaning would take away the pretext.