Amend @GeorgeMasonCC to require report of expenditures "from time-to-time", not annual
James Madison @JamesMadison316
#Move2 replace annual with "from time-to-time". Annual=too much & creates habit of doing nothing (like now) #Art1S9 - ow.ly/oNbjg

Mr. Madison proposed to strike out "annually" from the motion & insert "from time to time". which would enjoin the duty of frequent publications and leave enough to the discretion of the Legislature. Require too much and the difficulty will beget a habit of doing nothing. The articles of Confederation require half-yearly publications on this subject- A punctual compliance being often impossible, the practice has ceased altogether-