Add exception for inspection duties, as agreed to
Art 1. Sect. 10. (paragraph 2). “No State shall, without the consent of Congress lay imposts or duties on imports or exports; nor with such consent, but to the use of the Treasury of the U. States.” In consequence of the proviso moved by Col: Mason: and agreed to on the 13 of Sepr., this part of the section was laid aside in favor of the following substitute viz. “No State shall, without the consent of Congress, lay any imposts or duties on imports or exports, except what may be absolutely necessary for executing its Inspection laws; and the nett produce of all duties and imposts, laid by any State on imports or exports, shall be for the use of the Treasury of the U. S; and all such laws shall be subject to the revision and controul of the Congress” 1908 Putnam Sons Vol. Page # 379-380 ConText # 3977, 3978, 3980