Allow Congress to give appointment power to President
Art II. Sect. 2. (paragraph 2) To the end of this, Mr. GOVERNr. MORRIS moved to annex “but the Congress may by law vest the appointment of such inferior officers as they think proper, in the President alone, in the Courts of law, or in the heads of Departments.” 1908 Putnam Sons Vol. Page # 383 ConText # 4002, 4006